basic vr working
