+++ app/build.gradle
@@ -68,6 +68,7 @@ dependencies {
     implementation "androidx.hilt:hilt-lifecycle-viewmodel:1.0.0-alpha03"
     androidTestAnnotationProcessor "com.google.dagger:hilt-android-compiler:$hilt_version"
     compileOnly 'com.google.android.wearable:wearable:2.8.1'
+    implementation 'com.google.vr:sdk-panowidget:1.180.0'
 }
 
 tasks.withType(Test) {
