+++ app/src/main/AndroidManifest.xml
@@ -26,7 +26,7 @@
         -->
         <activity android:name=".scrolling.ad.AnnounceActivity" />
         <activity android:name=".scrolling.ScrollingActivity" />
-        <activity android:name=".vtour.VirtualTourActivity" />
+        <activity android:name=".virtualtour.VirtualTourActivity" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
